1.9.2 nifi version

+++ plc4j/integrations/apache-nifi/nifi-plc4x-processors/src/main/java/org/apache/plc4x/nifi/record/Plc4xReadResponseRecordSet.java
@@ -145,7 +145,7 @@ public class Plc4xReadResponseRecordSet implements RecordSet, Closeable {
             case BIT:
             	return RecordFieldType.BOOLEAN.getDataType();
             case DECIMAL:
-            	return RecordFieldType.DECIMAL.getDataType();
+            	return RecordFieldType.DOUBLE.getDataType();
             case BYTE:
             	return RecordFieldType.BYTE.getDataType();
             case CHAR:
